refactor: make `axvm-instructions` crate to separate out instructions
