+++ vm/src/arch/mod.rs
@@ -5,7 +5,7 @@ mod execution;
 /// Traits and wrappers to faciliate VM chip integration
 mod integration_api;
 
-pub mod instructions;
+pub use axvm_instructions as instructions;
 
 /// Testing framework
 #[cfg(test)]
